+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGKTM7DGKXSMKFKRPJE1ZTZ7"
+        "01KGKXJK5CP6P4PCF2MVZWX34Z"
     )
     r"""The unique identifier of the data source"""