+++ packages/orq-rc/src/orq_ai_sdk/models/createidentityop.py
@@ -137,7 +137,7 @@ class CreateIdentityResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-02-04T07:59:32.361Z")
+    updated: Optional[datetime] = parse_datetime("2026-02-04T08:51:04.128Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")