+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGKTM7DC0XTSZQQ1Q4WS6J12"
+        "01KGKXJK543H3ZBT68V8SY5QPB"
     )
     r"""The unique identifier of the data source"""