+++ packages/orq-rc/RELEASES.md
@@ -5239,3 +5239,13 @@ Based on:
 - [python v4.3.0-rc.22] packages/orq-rc
 ### Releases
 - [PyPI v4.3.0-rc.22] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.22 - packages/orq-rc
+
+## 2026-02-04 08:53:52
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.701.0 (2.803.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.3.0-rc.23] packages/orq-rc
+### Releases
+- [PyPI v4.3.0-rc.23] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.23 - packages/orq-rc